Add methods for interacting with links
Add a `$link` and `$links` methods for interacting with the links
themselves, as opposed to the resources in the links.

Added some basic tests.  I still need to update the documentation to
reflect these, however I want to make sure they are useful and
appropriate before spending the effort doing this.

Also, note the approach of having two methods: a `$link` and `$links`
method; I will probably add a `$rels` method in the future to act
similarly).